Common: merged room_file_base.cpp back into room_file.cpp
Certain Tools that work with CRM format will need to link all the related dependencies of the RoomData struct.
