+++ Solutions/Common.Lib/Common.Lib.vcxproj.filters
@@ -476,9 +476,6 @@
     <ClCompile Include="..\..\libsrc\allegro\src\math.c">
       <Filter>Library Sources\allegro</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Common\game\room_file_base.cpp">
-      <Filter>Source Files\game</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\Common\util\memorystream.cpp">
       <Filter>Source Files\util</Filter>
     </ClCompile>
