+++ Tools/CMakeLists.txt
@@ -30,14 +30,23 @@ set(TOOLS_COMMON_SOURCES
         ../Common/data/tra_file.h
         ../Common/debug/debugmanager.cpp
         ../Common/debug/debugmanager.h
-        ../Common/game/room_file_base.cpp
+        ../Common/game/customproperties.cpp
+        ../Common/game/customproperties.h
+        ../Common/game/interactions.cpp
+        ../Common/game/interactions.h
+        ../Common/game/room_file.cpp
+        ../Common/game/roomdata.cpp
         ../Common/gfx/bitmapdata.cpp
         ../Common/gfx/bitmapdata.h
-		../Common/gfx/image_bmp.cpp
+        ../Common/gfx/image_bmp.cpp
         ../Common/gfx/image_file.cpp
         ../Common/gfx/image_file.h
-		../Common/gfx/image_pcx.cpp
-		../Common/gfx/image_png.cpp
+        ../Common/gfx/image_pcx.cpp
+        ../Common/gfx/image_png.cpp
+        ../Common/script/cc_common.cpp
+        ../Common/script/cc_common.h
+        ../Common/script/cc_script.cpp
+        ../Common/script/cc_script.h
         ../Common/util/bufferedstream.cpp
         ../Common/util/bufferedstream.h
         ../Common/util/cmdlineopts.cpp
@@ -120,6 +129,7 @@ target_sources(libtools
         data/data_file_writer.h
         data/dialogscriptconv.cpp
         data/dialogscriptconv.h
+        data/cc_script_stubs.cpp
         data/game_utils.h
         data/game_utils.cpp
         data/include_utils.cpp
